little v for validation, renammed protected page
