+++ server/app/services/validateHomeStructure.js
